added cardDetails file
