+++ frontend/src/components/regions/AgileBoard/AgileCard/AgileCardActions/AgileCardActions.test.js
@@ -1,4 +1,4 @@
-import { showButtons } from "./utils";
+import { showButtons } from "../../../../../utils/cardButtons";
 
 import {
   READY,
